splitting up EDA and Train Test Split into 2 files
