--- data/processed/y_train.csv
@@ -1,2129 +1,0 @@
-quality
-5
-6
-6
-<<<<<<< HEAD
-5
-8
-5
-5
-5
-5
-5
-6
-5
-5
-7
-5
-6
-5
-5
-4
-5
-7
-6
-5
-5
-6
-5
-5
-6
-6
-5
-5
-5
-=======
-6
-6
-6
-7
-6
-6
->>>>>>> b040605b393c6209d5e528615f81d72e850e88bb
-4
-5
-6
-4
-6
-5
-5
-<<<<<<< HEAD
-5
-7
-6
-7
-5
-6
-5
-7
-5
-7
-6
-6
-5
-6
-6
-5
-5
-6
-7
-7
-7
-5
-5
-6
-7
-5
-5
-6
-4
-4
-4
-5
-5
-6
-7
-5
-7
-5
-5
-7
-7
-5
-5
-6
-5
-6
-4
-5
-5
-6
-5
-5
-6
-5
-6
-5
-5
-5
-7
-7
-6
-6
-6
-5
-5
-6
-6
-5
-6
-5
-6
-6
-4
-5
-5
-6
-8
-7
-5
-5
-5
-5
-6
-5
-5
-5
-6
-5
-6
-7
-7
-6
-5
-6
-5
-5
-5
-6
-5
-5
-6
-6
-6
-5
-5
-5
-5
-5
-7
-5
-7
-6
-6
-6
-5
-5
-5
-7
-7
-5
-5
-4
-6
-6
-6
-6
-6
-7
-6
-5
-7
-5
-6
-6
-5
-5
-6
-6
-8
-5
-6
-5
-5
-5
-6
-7
-5
-5
-6
-5
-5
-5
-6
-6
-5
-5
-6
-5
-5
-7
-5
-6
-6
-5
-5
-5
-6
-5
-6
-6
-6
-6
-7
-5
-6
-5
-7
-6
-6
-5
-5
-5
-5
-5
-4
-7
-6
-5
-5
-5
-5
-6
-8
-6
-6
-6
-3
-6
-6
-6
-5
-6
-7
-6
-6
-5
-8
-6
-6
-6
-6
-6
-5
-7
-7
-6
-7
-5
-6
-5
-5
-5
-6
-4
-5
-5
-4
-6
-7
-5
-5
-6
-6
-5
-5
-6
-7
-6
-6
-6
-5
-6
-5
-5
-7
-6
-7
-6
-5
-5
-5
-5
-8
-6
-6
-5
-5
-6
-5
-5
-6
-6
-7
-5
-6
-5
-7
-7
-5
-7
-6
-6
-5
-6
-6
-5
-6
-5
-5
-6
-6
-5
-5
-6
-5
-6
-5
-6
-7
-5
-6
-5
-7
-5
-5
-5
-5
-6
-6
-4
-6
-5
-7
-7
-6
-5
-6
-5
-5
-6
-5
-4
-6
-6
-6
-6
-7
-6
-6
-7
-5
-5
-5
-5
-5
-6
-6
-6
-6
-5
-6
-7
-7
-6
-5
-7
-5
-8
-5
-5
-7
-6
-5
-6
-6
-5
-6
-6
-5
-6
-6
-5
-5
-5
-6
-6
-6
-5
-6
-6
-6
-5
-7
-5
-5
-6
-5
-6
-5
-4
-6
-5
-6
-5
-5
-6
-5
-5
-6
-8
-5
-6
-6
-5
-5
-4
-6
-7
-7
-6
-5
-6
-5
-5
-7
-5
-5
-6
-6
-5
-6
-7
-5
-6
-5
-6
-5
-6
-5
-5
-5
-6
-6
-6
-6
-6
-7
-6
-7
-5
-5
-6
-=======
-7
-5
-6
-5
-7
->>>>>>> b040605b393c6209d5e528615f81d72e850e88bb
-6
-5
-3
-6
-<<<<<<< HEAD
-5
-6
-5
-5
-6
-6
-6
-6
-=======
-6
-6
-5
-5
-5
-6
-7
-6
-5
-7
-6
-7
-6
-7
-5
->>>>>>> b040605b393c6209d5e528615f81d72e850e88bb
-4
-6
-6
-6
-<<<<<<< HEAD
-6
-5
-5
-6
-6
-5
-5
-5
-4
-5
-6
-5
-5
-5
-5
-6
-6
-6
-6
-5
-6
-6
-5
-6
-5
-7
-7
-5
-5
-5
-5
-5
-8
-6
-7
-6
-5
-7
-5
-5
-6
-5
-6
-6
-5
-6
-5
-6
-6
-6
-7
-7
-5
-5
-5
-8
-6
-6
-7
-=======
-5
-5
-7
-5
-5
-8
->>>>>>> b040605b393c6209d5e528615f81d72e850e88bb
-6
-5
-5
-5
-7
-7
-6
-5
-6
-6
-<<<<<<< HEAD
-5
-6
-6
-7
-5
-7
-=======
->>>>>>> b040605b393c6209d5e528615f81d72e850e88bb
-5
-6
-6
-5
-<<<<<<< HEAD
-6
-5
-6
-5
-5
-5
-5
-5
-7
-6
-6
-4
-6
-=======
-7
-6
-5
-7
-6
-6
-4
-5
-5
-6
-7
-6
-6
-6
-6
-8
-6
->>>>>>> b040605b393c6209d5e528615f81d72e850e88bb
-6
-5
-6
-5
-6
-5
-6
-6
-<<<<<<< HEAD
-3
-5
-5
-6
-6
-6
-6
-8
-5
-5
-=======
-6
-5
-7
-5
-5
-5
-6
-5
-5
-6
-5
-6
-5
-6
-6
-7
-7
-5
-7
-6
-6
-6
-6
-5
-5
-6
-5
-5
-6
-6
-6
-6
-6
-5
-5
-5
-7
-5
-5
-7
-7
-7
-6
-7
-6
-5
-5
-5
-6
-6
-5
-5
-6
-7
-6
-6
-5
-5
-6
-6
-6
-6
-5
-5
-6
-5
-5
-6
-5
-7
-6
-4
-6
-6
-6
-5
-5
-5
-3
-5
-7
-6
-5
-6
-5
-7
-5
-5
-6
-5
-5
-5
-5
-5
-5
-5
-5
-5
-6
-5
-6
-6
-5
-6
-5
-7
-6
-5
-7
-5
-6
-5
-6
-6
-6
-5
-6
-6
->>>>>>> b040605b393c6209d5e528615f81d72e850e88bb
-5
-5
-5
-5
-5
-<<<<<<< HEAD
-6
-6
-=======
->>>>>>> b040605b393c6209d5e528615f81d72e850e88bb
-5
-6
-5
-<<<<<<< HEAD
-7
-5
-5
-5
-5
-=======
-6
-6
-4
-5
-5
-6
-6
-5
-4
-6
-6
->>>>>>> b040605b393c6209d5e528615f81d72e850e88bb
-6
-7
-5
-5
-<<<<<<< HEAD
-5
-7
-5
-5
-8
-5
-5
-=======
->>>>>>> b040605b393c6209d5e528615f81d72e850e88bb
-5
-4
-5
-6
-<<<<<<< HEAD
-5
-6
-5
-5
-6
-=======
-6
-6
-6
-6
-6
-6
-5
-5
-6
-5
-5
->>>>>>> b040605b393c6209d5e528615f81d72e850e88bb
-3
-6
-5
-5
-5
-5
-6
-<<<<<<< HEAD
-5
-6
-6
-5
-5
-5
-7
-5
-6
-5
-7
-6
-5
-5
-5
-5
-4
-5
-4
-=======
-7
-5
-5
-7
-5
-7
-5
-5
-7
-4
-6
-6
-6
-5
-7
-7
-5
->>>>>>> b040605b393c6209d5e528615f81d72e850e88bb
-6
-6
-5
-5
-<<<<<<< HEAD
-6
-5
-7
-6
-3
-5
-5
-6
-5
-6
-5
-6
-5
-6
-7
-6
-7
-6
-4
-=======
-7
-6
-8
-6
-7
-5
-6
-6
-5
-5
-5
-5
-7
-5
-5
-6
-5
-5
-5
-5
-5
-5
-7
-7
-5
-6
-6
-6
-6
-7
-6
-5
-6
-5
-6
-5
-6
-6
-7
-6
-7
-7
-5
-5
-7
-6
-5
-6
-6
-6
-5
-6
-5
-5
-6
-6
-7
-6
-6
-6
-5
-6
-4
-5
-5
-6
-7
-6
-5
-6
-5
-5
-5
-6
-6
-5
-5
-5
-6
-6
-7
-7
-5
-6
-5
-6
-5
-7
-6
-5
-6
-6
-5
-5
-5
-6
-6
-6
-8
-6
-5
-7
-5
-6
-5
-6
-5
-5
-5
-4
-7
-5
-5
-6
-6
-5
-5
-6
-7
-5
-6
-5
-6
-5
-7
-5
-5
-6
-6
-5
-6
-5
-5
-7
-6
-5
-7
-6
-8
-6
-7
-6
-5
-6
-6
-7
-5
-5
-5
-7
-5
-5
-5
-7
-5
-5
-5
-5
-5
-6
-6
-5
-6
-7
-5
-6
-6
-6
-6
-6
-6
-6
-5
-5
-5
-5
-5
-6
-7
-5
-6
-5
-5
-6
-5
-5
-5
-5
-6
-5
-5
-5
-7
-5
-7
-7
-5
-7
-5
-6
-5
-5
-5
-6
-6
-5
-6
-5
-5
-6
-6
-6
-5
-5
-4
-6
-7
-8
-5
-5
-6
-5
-5
-5
-6
-6
-5
-5
-7
-6
-6
-6
-6
-7
-5
-7
-4
-4
-6
-5
-5
-6
-5
-5
-6
-5
-5
-6
-5
-6
-6
-6
-3
-5
-5
-5
-6
-5
-6
-6
-6
-4
-6
-6
-5
-5
-5
-6
-6
->>>>>>> b040605b393c6209d5e528615f81d72e850e88bb
-4
-5
-5
-6
-<<<<<<< HEAD
-6
-7
-6
-5
-7
-5
-6
-5
-5
-6
-5
-5
-5
-5
-6
-6
-5
-7
-6
-5
-6
-6
-5
-5
-5
-7
-6
-=======
->>>>>>> b040605b393c6209d5e528615f81d72e850e88bb
-5
-6
-7
-7
-5
-5
-5
-6
-5
-7
-5
-<<<<<<< HEAD
-5
-5
-5
-6
-6
-6
-5
-5
-6
-6
-6
-5
-4
-5
-5
-5
-5
-5
-5
-6
-7
-5
-7
-5
-6
-6
-=======
-5
-5
-6
-6
-6
-5
-5
-7
-6
-5
-5
-5
-6
-6
-5
-7
-6
-6
-5
-5
-5
-5
-7
-5
-5
-5
-5
-5
-5
-6
-6
-5
-5
-6
-5
-5
-4
-4
->>>>>>> b040605b393c6209d5e528615f81d72e850e88bb
-6
-6
-6
-5
-<<<<<<< HEAD
-6
-4
-5
-6
-6
-6
-5
-6
-6
-5
-5
-4
-5
-4
-5
-6
-6
-6
-5
-6
-7
-7
-5
-6
-5
-6
-7
-6
-6
-7
-6
-6
-5
-5
-5
-5
-6
-4
-5
-6
-6
-5
-6
-8
-5
-6
-6
-6
-5
-5
-6
-5
-4
-5
-7
-6
-5
-5
-5
-5
-5
-7
-5
-6
-7
-6
-7
-5
-7
-5
-5
-6
-6
-6
-7
-6
-6
-7
-6
-5
-7
-5
-4
-5
-6
-5
-7
-6
-6
-6
-6
-5
-5
-7
-6
-5
-5
-5
-6
-7
-5
-5
-6
-5
-6
-6
-6
-6
-5
-5
-5
-5
-6
-4
-7
-7
-6
-5
-6
-6
-6
-5
-7
-5
-6
-5
-6
-6
-6
-5
-5
-6
-6
-6
-5
-5
-5
-5
-5
-6
-6
-5
-7
-6
-7
-6
-5
-6
-6
-7
-6
-5
-6
-6
-4
-6
-5
-6
-5
-6
-5
-5
-6
-5
-6
-6
-6
-6
-6
-8
-6
-5
-5
-7
-5
-6
-5
-6
-5
-7
-5
-5
-6
-6
-6
-5
-6
-6
-5
-5
-5
-5
-6
-5
-5
-5
-=======
-5
-6
-5
-7
-4
-6
-6
-3
-5
-6
-6
-4
-3
-5
-7
-8
-6
-6
-5
-6
-5
-6
-7
-6
-4
-5
-8
-5
-6
-6
-5
-5
-6
-5
-6
-5
-5
-5
-5
-7
->>>>>>> b040605b393c6209d5e528615f81d72e850e88bb
-6
-5
-5
-6
-6
-5
-<<<<<<< HEAD
-4
-5
-7
-7
-6
-7
-5
-=======
-6
->>>>>>> b040605b393c6209d5e528615f81d72e850e88bb
-6
-5
-6
-5
-5
-<<<<<<< HEAD
-5
-6
-6
-5
-5
-6
-6
-6
-6
-6
-5
-5
-5
-6
-6
-6
-5
-5
-6
-6
-6
-6
-5
-6
-6
-6
-8
-6
-6
-5
-5
-6
-7
-6
-6
-3
-6
-4
-=======
-6
-6
-6
-3
-6
-6
-7
-5
-6
-6
-6
-6
->>>>>>> b040605b393c6209d5e528615f81d72e850e88bb
-5
-5
-5
-5
-<<<<<<< HEAD
-=======
-5
-6
-7
-5
-5
-6
-5
-5
-5
-6
-5
-6
-5
-5
-7
-6
-8
-5
-5
-5
-6
-6
-7
-5
-4
-6
-5
-5
-5
-5
-6
-7
-7
-7
-5
-6
-5
-7
-6
-7
-5
-5
-5
-7
-5
-4
-7
-5
-5
-7
-6
-4
-5
-5
-5
-5
-4
-5
-6
-5
-5
-6
-5
-6
-6
-7
-6
-6
-7
-5
-5
-5
-5
-7
-6
-5
-7
-6
-5
-6
-7
-5
-5
-7
-5
-5
-6
-5
-7
-6
-6
-5
-6
-6
-6
-5
-5
-5
-5
-6
-6
-6
-5
-6
-5
-5
-5
-6
-6
-6
-6
-5
-6
-5
-7
-6
-5
-5
-4
-6
-6
-6
-5
-6
-6
-5
-6
-7
-6
-5
-6
-7
-8
-5
-6
-5
-6
-7
-6
-6
-6
-4
-6
-5
-4
-5
-4
-7
-7
-5
-5
-5
-6
-5
-5
-5
-5
-5
-5
-6
-6
-6
-5
-6
-5
-5
-6
-5
-7
-6
-6
-5
-5
-6
-7
-7
-6
-7
-8
-4
-5
-5
-5
-6
-6
-8
-6
-7
-5
-6
-6
-6
-5
-5
-5
-5
-6
-7
-5
-5
-5
-5
-4
-7
-5
-6
-5
-4
-6
-5
-4
-5
-6
-5
-5
-6
-6
-6
-6
-6
-5
-5
-6
-5
-5
-6
-5
-5
-6
-6
-5
-6
-6
-6
-6
-5
-6
-5
-5
-5
-6
-4
-7
-7
-5
-5
-7
-5
-6
-5
-5
-5
-5
-4
-6
-5
-6
-7
-7
-6
-6
-6
-5
-6
-5
-4
-6
-7
-7
-6
-6
-6
->>>>>>> b040605b393c6209d5e528615f81d72e850e88bb
-6
-6
-7
-5
-6
-6
-5
-6
-5
-5
-5
-7
-6
-5
-5
-6
-7
-6
-5
-7
-7
-7
-5
-6
-7
-6
-4
-6
-6
-6
-7
-5
-5
-5
-7
-5
-7
-5
-6
-7
-6
-5
-5
-6
-5
-5
-6
-5
-7
-6
-5
-5
-6
-6
-6
-5
-5
-6
-4
-6
-7
-5
-6
-4
-5
-6
-7
-5
-5
-5
-6
-4
-5
-6
-6
-5
-5
-5
-4
-6
-6
-5
-5
-5
-6
-6
-5
-5
-4
-4
-7
-6
-4
-5
-6
-5
-3
-6
-6
-<<<<<<< HEAD
-6
-6
-6
-5
-5
-5
-6
-6
-7
-5
-7
-6
-6
-5
-6
-5
-5
-5
-4
-6
-6
-5
-5
-4
-=======
-4
-6
-6
-7
-6
-5
-5
-8
-5
-5
-5
-5
-5
->>>>>>> b040605b393c6209d5e528615f81d72e850e88bb
-6
-5
-7
-6
-6
-5
-5
-5
-5
-5
-6
-5
-6
-6
-6
-6
-6
-5
-6
-5
-6
-5
-4
-6
-4
-6
-5
-6
-5
-7
-5
-7
-5
-<<<<<<< HEAD
-6
-7
-5
-6
-=======
->>>>>>> b040605b393c6209d5e528615f81d72e850e88bb
-5
-7
-6
-6
-<<<<<<< HEAD
-5
-6
-6
-=======
->>>>>>> b040605b393c6209d5e528615f81d72e850e88bb
-6
-5
-4
-6
-5
-6
-6
-7
-<<<<<<< HEAD
-4
-=======
-6
-7
-5
-5
-5
-5
->>>>>>> b040605b393c6209d5e528615f81d72e850e88bb
-6
-6
-5
-6
-6
-<<<<<<< HEAD
-6
-5
-4
-5
-7
-=======
-7
-5
-5
-7
-7
-7
-6
-6
-6
-5
-6
-3
-5
-5
->>>>>>> b040605b393c6209d5e528615f81d72e850e88bb
