--- results/tables/best_params.csv
@@ -1,2 +1,0 @@
-svc__C,svc__class_weight,svc__decision_function_shape,svc__gamma,best_score
-3.907967156822881,,ovo,0.008629132190071852,0.5694245973026677
