--- results/tables/confusion_matrix_summary.csv
@@ -1,5 +1,0 @@
-,3,4,5,6,7,8
-True Negative,270,262,109,107,240,268
-False Positive,0,0,53,45,6,0
-False Negative,2,10,19,47,22,4
-True Positive,0,0,91,73,4,0
