--- results/tables/describe_table.csv
@@ -1,19 +1,0 @@
-,fixed acidity,volatile acidity,citric acid,residual sugar,chlorides,free sulfur dioxide,total sulfur dioxide,density,pH,sulphates,alcohol
-count,1087.0,1087.0,1087.0,1087.0,1087.0,1087.0,1087.0,1087.0,1087.0,1087.0,1087.0
-<<<<<<< HEAD
-mean,8.348942042318306,0.5277736890524379,0.2787580496780129,2.5425482980680774,0.08811131554737811,16.03863845446182,47.64995400183992,0.9967230174793007,3.3053909843606255,0.6599172033118675,10.431600735970562
-std,1.7648126803777153,0.18310565720698913,0.19618794163028544,1.367262135427783,0.050714576409770375,10.51513323138459,33.76872725771016,0.0018708735569719012,0.1552797298564071,0.1741474697051162,1.0814567082763253
-min,4.6,0.16,0.0,0.9,0.012,1.0,6.0,0.99007,2.74,0.37,8.5
-25%,7.1,0.39,0.1,1.9,0.07,7.5,22.0,0.9956,3.2,0.55,9.5
-50%,7.9,0.52,0.27,2.2,0.079,14.0,38.0,0.99674,3.3,0.62,10.2
-75%,9.3,0.635,0.435,2.6,0.091,22.0,64.0,0.997835,3.39,0.73,11.1
-max,15.9,1.58,1.0,15.5,0.611,72.0,289.0,1.0032,4.01,2.0,14.9
-=======
-mean,8.271297148114076,0.5321619135234591,0.26917203311867527,2.535970561177553,0.0882161913523459,15.701011959521619,46.10533578656854,0.996678840846366,3.3102759889604414,0.6547930082796688,10.436829193498927
-std,1.7081174604713758,0.18183272396828717,0.1968206193387154,1.4059098333597764,0.04934127961330793,10.450348663495538,33.522362578821394,0.0018475896246839543,0.15340253023655095,0.16557986846961564,1.0776486085878827
-min,4.6,0.12,0.0,0.9,0.012,1.0,6.0,0.99007,2.74,0.33,8.5
-25%,7.1,0.4,0.09,1.9,0.07,7.0,21.0,0.995575,3.21,0.55,9.5
-50%,7.9,0.52,0.26,2.2,0.079,13.0,37.0,0.99668,3.31,0.62,10.1
-75%,9.2,0.64,0.42,2.6,0.09,21.0,62.0,0.9978,3.4,0.72,11.1
-max,15.9,1.58,1.0,15.5,0.611,72.0,289.0,1.00369,4.01,2.0,14.9
->>>>>>> b040605b393c6209d5e528615f81d72e850e88bb
