--- results/tables/initial_model_scores.csv
@@ -1,7 +1,0 @@
-,fit_time,score_time,test_score,train_score
-dummy,0.003 (+/- 0.001),0.003 (+/- 0.001),0.430 (+/- 0.003),0.430 (+/- 0.001)
-decision tree,0.008 (+/- 0.001),0.003 (+/- 0.000),0.499 (+/- 0.035),1.000 (+/- 0.000)
-kNN,0.004 (+/- 0.000),0.014 (+/- 0.001),0.544 (+/- 0.027),0.678 (+/- 0.005)
-RBF SVM,0.038 (+/- 0.002),0.011 (+/- 0.001),0.584 (+/- 0.026),0.662 (+/- 0.008)
-naive bayes,0.004 (+/- 0.000),0.003 (+/- 0.001),0.541 (+/- 0.038),0.557 (+/- 0.012)
-log reg,0.015 (+/- 0.001),0.004 (+/- 0.000),0.568 (+/- 0.030),0.585 (+/- 0.011)
