--- results/tables/test_accuracy.csv
@@ -1,2 +1,0 @@
-accuracy
-0.6176470588235294
